+++ ntex-net/src/lib.rs
@@ -8,7 +8,7 @@
     clippy::missing_errors_doc,
     clippy::missing_panics_doc
 )]
-use std::{io, net, net::SocketAddr, panic};
+use std::{any::Any, io, net, net::SocketAddr, panic};
 
 use ntex_io::Io;
 use ntex_rt::{BlockFuture, Driver, Runner};
@@ -108,29 +108,21 @@ pub struct DefaultRuntime;
 
 impl Runner for DefaultRuntime {
     #[allow(unused_variables, clippy::too_many_lines)]
-    fn block_on(&self, fut: BlockFuture) {
+    fn block_on(&self, fut: BlockFuture) -> Result<(), Box<dyn Any + Send>> {
         #[cfg(feature = "tokio")]
         {
             let driver: Box<dyn Reactor> = Box::new(self::tokio::TokioDriver);
 
-            CURRENT_DRIVER.set(&driver, || {
-                crate::tokio::block_on(fut);
-                unsafe {
-                    ntex_rt::remove_all_items();
-                }
-            });
+            CURRENT_DRIVER.set(&driver, || crate::tokio::block_on(fut));
+            Ok(())
         }
 
         #[cfg(all(feature = "compio", not(feature = "tokio")))]
         {
             let driver: Box<dyn Reactor> = Box::new(self::compio::CompioDriver);
 
-            CURRENT_DRIVER.set(&driver, || {
-                crate::compio::block_on(fut);
-                unsafe {
-                    ntex_rt::remove_all_items();
-                }
-            });
+            CURRENT_DRIVER.set(&driver, || crate::compio::block_on(fut));
+            Ok(())
         }
 
         #[cfg(all(windows, not(feature = "tokio"), not(feature = "compio")))]
@@ -139,19 +131,11 @@ impl Runner for DefaultRuntime {
             let driver: Box<dyn Reactor> = Box::new(driver);
 
             CURRENT_DRIVER.set(&driver, || {
-                let res = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+                panic::catch_unwind(panic::AssertUnwindSafe(|| {
                     let rt = ntex_rt::Runtime::new(driver.handle());
                     rt.block_on(fut, &*driver);
-                }));
-                unsafe {
-                    if let Err(err) = res {
-                        ntex_rt::remove_all_items();
-                        panic::resume_unwind(err);
-                    } else {
-                        ntex_rt::remove_all_items();
-                    }
-                }
-            });
+                }))
+            })
         }
 
         #[cfg(all(unix, not(feature = "tokio"), not(feature = "compio")))]
@@ -163,19 +147,11 @@ impl Runner for DefaultRuntime {
                 let driver: Box<dyn Reactor> = Box::new(driver);
 
                 CURRENT_DRIVER.set(&driver, || {
-                    let res = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+                    panic::catch_unwind(panic::AssertUnwindSafe(|| {
                         let rt = ntex_rt::Runtime::new(driver.handle());
                         rt.block_on(fut, &*driver);
-                    }));
-                    unsafe {
-                        if let Err(err) = res {
-                            ntex_rt::remove_all_items();
-                            panic::resume_unwind(err);
-                        } else {
-                            ntex_rt::remove_all_items();
-                        }
-                    }
-                });
+                    }))
+                })
             }
 
             #[cfg(all(target_os = "linux", feature = "neon-uring"))]
@@ -185,19 +161,11 @@ impl Runner for DefaultRuntime {
                 let driver: Box<dyn Reactor> = Box::new(driver);
 
                 CURRENT_DRIVER.set(&driver, || {
-                    let res = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+                    panic::catch_unwind(panic::AssertUnwindSafe(|| {
                         let rt = ntex_rt::Runtime::new(driver.handle());
                         rt.block_on(fut, &*driver);
-                    }));
-                    unsafe {
-                        if let Err(err) = res {
-                            ntex_rt::remove_all_items();
-                            panic::resume_unwind(err);
-                        } else {
-                            ntex_rt::remove_all_items();
-                        }
-                    }
-                });
+                    }))
+                })
             }
 
             #[cfg(all(not(feature = "neon-uring"), not(feature = "neon-polling")))]
@@ -218,19 +186,11 @@ impl Runner for DefaultRuntime {
                 );
 
                 CURRENT_DRIVER.set(&driver, || {
-                    let res = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+                    panic::catch_unwind(panic::AssertUnwindSafe(|| {
                         let rt = ntex_rt::Runtime::new(driver.handle());
                         rt.block_on(fut, &*driver);
-                    }));
-                    unsafe {
-                        if let Err(err) = res {
-                            ntex_rt::remove_all_items();
-                            panic::resume_unwind(err);
-                        } else {
-                            ntex_rt::remove_all_items();
-                        }
-                    }
-                });
+                    }))
+                })
             }
         }
     }