+++ ntex/src/web/server.rs
@@ -150,22 +150,31 @@ where
     }
 
     #[must_use]
+    /// Stops the server when one of the workers panics.
+    ///
+    /// By default, "stop on panic" is disabled.
+    pub fn stop_on_panic(mut self) -> Self {
+        self.builder = self.builder.stop_on_panic();
+        self
+    }
+
+    #[must_use]
     /// Disable signal handling.
     ///
-    /// By default signal handling is enabled.
+    /// By default, signal handling is enabled.
     pub fn disable_signals(mut self) -> Self {
         self.builder = self.builder.disable_signals();
         self
     }
 
     #[must_use]
-    /// Timeout for graceful workers shutdown.
+    /// Timeout for graceful worker shutdown.
     ///
     /// After receiving a stop signal, workers have this much time to finish
-    /// serving requests. Workers still alive after the timeout are force
-    /// dropped.
+    /// serving requests. Workers that are still alive after the timeout are
+    /// forcefully dropped.
     ///
-    /// By default shutdown timeout sets to 30 seconds.
+    /// By default, the shutdown timeout is set to 30 seconds.
     pub fn shutdown_timeout(mut self, sec: Seconds) -> Self {
         self.builder = self.builder.shutdown_timeout(sec);
         self
@@ -174,9 +183,19 @@ where
     #[must_use]
     /// Enable cpu affinity.
     ///
-    /// By default affinity is disabled.
+    /// By default, affinity is disabled.
     pub fn enable_affinity(mut self) -> Self {
         self.builder = self.builder.enable_affinity();
+        self
+    }
+
+    #[must_use]
+    /// Graceful shutdown.
+    ///
+    /// Gracefully shuts down on SIGTERM, SIGSEGV, or SIGQUIT.
+    /// By default, graceful shutdown is disabled.
+    pub fn graceful_shutdown(mut self) -> Self {
+        self.builder = self.builder.graceful_shutdown();
         self
     }